+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.1.3 - Under development
 -------------------------
 
-* Adds a ``type`` field to the ``Article`` model
+* Adds a ``type`` field to the ``Article`` model and update the ``import_journals_from_fedora`` command
 * Remove old ``get_absolute_url`` methods
 
 0.1.2